Migrated CommitLogger completely.
