+++ plug-ins/git/src/main/java/eu/sqooss/plugins/tds/git/GitCommitLog.java
@@ -30,12 +30,7 @@
 
 package eu.sqooss.plugins.tds.git;
 
-import java.util.Iterator;
-import java.util.LinkedList;
-import java.util.List;
-
 import eu.sqooss.plugins.tds.scm.SCMCommitLog;
-import eu.sqooss.service.tds.Revision;
 
 /**
  * The Git implementation of the CommitLog interface.
@@ -45,33 +40,7 @@ import eu.sqooss.service.tds.Revision;
  */
 public class GitCommitLog extends SCMCommitLog {
 
-    private LinkedList<Revision> entries;
-    
     public GitCommitLog() {
-        entries = new LinkedList<Revision>();
-    }
-    
-    public List<Revision> entries() {
-        return entries;
-    }
-
-    @Override
-    public Iterator<Revision> iterator() {
-        return entries.iterator();
-    }
-
-    @Override
-    public Revision first() {
-        return entries.getFirst();
-    }
-
-    @Override
-    public Revision last() {
-        return entries.getLast();
-    }
-
-    @Override
-    public int size() {
-        return entries.size();
+    	super();
     }
 }
